requirements.txt
confluent-kafka>=2.3.0
boto3>=1.34.0
snowflake-connector-python>=3.10.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
watchdog>=4.0.0
streamlit>=1.31.1
plotly>=5.18.0
altair>=5.2.0
